Add basic support for defining the bounds of a range
Currently, this is done with a hard-coded index type of Int. And, the only check for out of bounds is on update. Index out of bounds on reference can be checked later?

Also, there is no propogation of index bounds for formulas. I think it should be possible for a range with one bounds to reference a range with a different bounds. But, I could be wrong.