Add support for ranges registering with the Sheet trait
Registering with the sheet has the problem that different ranges have different types (literal types too). We need to register the typed ranges into a symbol table that has less typing so that later phases have access to the ranges in a single collection (which currently doesn't allow for heterogeneous types).